+++ CHANGELOG.md
@@ -3,6 +3,13 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 ### Enhancements
+
+### Bug fixes
+
+### Deprecations
+
+## [2.0.2] - 2018-10-02
+### Enhancements
 - Add finger print feature for Android emulators [#13](https://github.com/appium/ruby_lib_core/issues/13)
 - Add `keyboard_shown?` and `context=` as aliases of `is_keyboard_shown` and `set_contex`